Code consistency fixes
changed all uses of MathHelper.PositiveOrNegativeOne to be stored as ints

+++ Assets/Scripts/Game/Projectiles/Enemy bullet systems/Aquarius/AquariusBulletSystem2.cs
@@ -23,7 +23,7 @@ public class AquariusBulletSystem2 : EnemyShooter<EnemyBullet>
 
         while (enabled)
         {
-            float d = PositiveOrNegativeOne;
+            int d = PositiveOrNegativeOne;
 
             for (int i = 1; i <= WaveCount; i++)
             {
